+++ src/crypto/test/crypto.cpp
@@ -890,8 +890,9 @@ TEST_CASE("AES-GCM convenience functions")
 {
   EntropyPtr entropy = get_entropy();
   std::vector<uint8_t> key = entropy->random(GCM_DEFAULT_KEY_SIZE);
-  auto encrypted = aes_gcm_encrypt(key, contents);
-  auto decrypted = aes_gcm_decrypt(key, encrypted);
+  auto iv = entropy->random(iv_size);
+  auto encrypted = aes_gcm_encrypt(key, contents, iv);
+  auto decrypted = aes_gcm_decrypt(key, encrypted, iv);
   REQUIRE(decrypted == contents);
 }
 
@@ -1336,10 +1337,13 @@ TEST_CASE("Sign and verify a chain with an intermediate and different subjects")
 
 TEST_CASE("Decrypt should validate integrity")
 {
-  auto key = get_entropy()->random(16);
+  auto entropy = ccf::crypto::get_entropy();
+  auto key = entropy->random(16);
+  auto iv = entropy->random(ccf::crypto::iv_size);
   std::vector<uint8_t> expected_plaintext = {0xde, 0xad, 0xbe, 0xef};
-  auto ciphertext = ccf::crypto::aes_gcm_encrypt(key, expected_plaintext);
-  auto decrypted_plaintext = ccf::crypto::aes_gcm_decrypt(key, ciphertext);
+  auto ciphertext = ccf::crypto::aes_gcm_encrypt(key, expected_plaintext, iv);
+  auto decrypted_plaintext =
+    ccf::crypto::aes_gcm_decrypt(key, ciphertext, iv);
 
   CHECK_EQ(expected_plaintext, decrypted_plaintext);
 
@@ -1348,7 +1352,7 @@ TEST_CASE("Decrypt should validate integrity")
   broken_ciphertext[ciphertext.size() / 2] =
     ~broken_ciphertext[ciphertext.size() / 2];
 
-  CHECK_THROWS(ccf::crypto::aes_gcm_decrypt(key, broken_ciphertext));
+  CHECK_THROWS(ccf::crypto::aes_gcm_decrypt(key, broken_ciphertext, iv));
 }
 
 TEST_CASE("Do not trust non-ca certs")