+++ src/pal/test/snp_ioctl_test.cpp
@@ -42,10 +42,11 @@ TEST_CASE("SNP derive key")
     ccf::ds::to_hex(key1->get_raw()), ccf::ds::to_hex(key2->get_raw()));
 
   std::vector<uint8_t> expected_plaintext = {0xde, 0xad, 0xbe, 0xef};
+  auto iv = ccf::crypto::get_entropy()->random(ccf::crypto::iv_size);
   auto ciphertext =
-    ccf::crypto::aes_gcm_encrypt(key1->get_raw(), expected_plaintext);
+    ccf::crypto::aes_gcm_encrypt(key1->get_raw(), expected_plaintext, iv);
   auto decrypted_plaintext =
-    ccf::crypto::aes_gcm_decrypt(key2->get_raw(), ciphertext);
+    ccf::crypto::aes_gcm_decrypt(key2->get_raw(), ciphertext, iv);
 
   CHECK_EQ(
     ccf::ds::to_hex(expected_plaintext), ccf::ds::to_hex(decrypted_plaintext));
@@ -63,13 +64,14 @@ TEST_CASE("SNP derived keys with different TCBs should be different")
   CHECK_NE(ccf::ds::to_hex(key1->get_raw()), ccf::ds::to_hex(key2->get_raw()));
 
   std::vector<uint8_t> expected_plaintext = {0xde, 0xad, 0xbe, 0xef};
+  auto iv = ccf::crypto::get_entropy()->random(ccf::crypto::iv_size);
   bool threw = false;
   try
   {
     auto ciphertext =
-      ccf::crypto::aes_gcm_encrypt(key1->get_raw(), expected_plaintext);
+      ccf::crypto::aes_gcm_encrypt(key1->get_raw(), expected_plaintext, iv);
     auto decrypted_plaintext =
-      ccf::crypto::aes_gcm_decrypt(key2->get_raw(), ciphertext);
+      ccf::crypto::aes_gcm_decrypt(key2->get_raw(), ciphertext, iv);
   }
   catch (std::runtime_error& e)
   {